refactor: features/deck/componentsにカードUIコンポーネントを移行（TASK-0070）
CardUI、HandDisplay、DraggableCardUIをsrc/features/deck/components/に移動し、
Feature-Based Architecture準拠のディレクトリ構造を実現。

変更点:
- src/features/deck/components/にCardUI、HandDisplay、DraggableCardUIを新規作成
- BaseComponentを@shared/componentsから継承するように更新
- 旧パス(presentation/ui/components/)から後方互換性のため再エクスポート設定
- インポートテストを追加
