+++ atelier-guild-rank/src/features/deck/components/index.ts
@@ -1,0 +1,16 @@
+/**
+ * Deck Components Module
+ * デッキ機能関連UIコンポーネントの公開エクスポート
+ *
+ * TASK-0070: features/deck/components作成
+ */
+
+// CardUI
+export type { CardUIConfig } from './CardUI';
+export { CardUI } from './CardUI';
+// DraggableCardUI
+export type { DraggableCardConfig } from './DraggableCardUI';
+export { DraggableCardUI } from './DraggableCardUI';
+// HandDisplay
+export type { HandDisplayConfig } from './HandDisplay';
+export { HandDisplay } from './HandDisplay';
