+++ atelier-guild-rank/src/features/deck/index.ts
@@ -1,0 +1,15 @@
+/**
+ * Deck Feature Module
+ * デッキ機能の公開エクスポート
+ *
+ * TASK-0069: features/deck/services作成
+ * TASK-0070: features/deck/components作成
+ */
+
+// Components
+export type { CardUIConfig, DraggableCardConfig, HandDisplayConfig } from './components';
+export { CardUI, DraggableCardUI, HandDisplay } from './components';
+export { draw } from './services/draw';
+export { playCard } from './services/play-card';
+// Services
+export { shuffle } from './services/shuffle';
